hospital_flask/requirements.txt
flask
mysql-connector-python
werkzeug
flask-session
